refactor(web): split modules/fizruk/pages/Body.tsx into per-component files
